==> ML_SELEX/step_03_assignNeighborhoods.ipynb <==
# Inputs:
local_kings_file = 'local_kings.csv'
dataset_file = 'dataset.csv'

# Outputs: 
# Outputs overwrite the dataset file which will now just be one columns bigger

import pandas as pd
import numpy as np
import random
from tqdm import tqdm
from polyleven import levenshtein

def assignLK(ix):
    
    if ix in df_lk.index:
        return ix
    
    sx = df.loc[ix, 'seq']
    
    df_lk['d'] = [levenshtein(sx, lk) for lk in df_lk['seq']]
    min_dist = df_lk['d'].min()
    candidates = df_lk[df_lk['d'] == min_dist]
    if len(candidates) == 1:
        return candidates.iloc[0].name
    else:
        return random.choice(candidates.index.tolist())

df = pd.read_csv(dataset_file, index_col=0)
df_lk = pd.read_csv(local_kings_file, index_col=0)

df['lk'] = [assignLK(x) for x in df.index]

df.head(20)

df.to_csv(dataset_file)
