==> ML_SELEX/step_04_parallel_getSingleMutants.ipynb <==
import pandas as pd
from polyleven import levenshtein
from multiprocessing import Pool
from functools import partial
import os
import pickle
import sys

def compute_sample_distances(i, seqs, dist_function):
    results = []
    for j in range(i + 1, len(seqs)):
        edit_dist = dist_function(seqs[i], seqs[j], 1)
        if edit_dist == 1:
            results.append(frozenset([i, j]))
    return results
            
def compute_dists(dist_function):
    partial_sample_dists = partial(compute_sample_distances, dist_function=dist_function, seqs=seqs)
    with Pool() as p:
        results_all = p.map(partial_sample_dists, list(range(len(seqs))))
    return results_all

input_file = 'dataset.csv'
output_file = 'single_mutants.pkl'
lk_index = 0

df = pd.read_csv(input_file, index_col=0)

df

local_king_indexes = sorted(df['lk'].unique().tolist())
lk = local_king_indexes[lk_index]

df_lk = df[df['lk'] == lk].head(5000) # !! This .head(5000) needs to be deleted in the real run
seqs = df_lk['seq'].values
ids = df_lk.index

single_mutants = compute_dists(levenshtein)

with open(output_file, 'wb') as f:
    pickle.dump(single_mutants, f)
