==> ML_SELEX/step_06_making_splits_n_encoding.ipynb <==
import pandas as pd
import random

df = pd.read_csv('paired_dataset_complete.csv', names=['i_b', 'i_w', 'seq_b', 'seq_w', 'count_b', 'count_w', 'deltaC', 'n'])

def makeLocalSplits(df, f):
    '''
    Here the input dataframe is subset based on neighborhood. Random sample (fraction f of the neighborhood size) is used for training
    and the rest is split equally into test and val
    '''
    
    split = {'train':[], 'val':[], 'test':[]}
    
    for lk in df['n'].unique():
        
        neighborhood_indexes = set(df[df['n'] == lk].index)
        
        train = set([random.choice(tuple(neighborhood_indexes)) for x in range(int(len(neighborhood_indexes)*f))])
        valtest = neighborhood_indexes - train
        val = set([random.choice(tuple(valtest)) for x in range(int(len(valtest)/2))])
        test = valtest - val
                  
        split['train'].extend(list(train))
        split['val'].extend(list(val))
        split['test'].extend(list(test))
                  
    return split

def makeGlobalSplits(df, f):
    '''
    TODO
    '''
    pass

fractions = [0.5, 0.33, 0.1, 0.01]

for f in fractions:
    
    local_splits = makeLocalSplits(df, f)
    
    df.loc[local_splits['train']].to_csv('splits/local_train_' + str(f).replace('.', '_') + '.csv')
    df.loc[local_splits['val']].to_csv('splits/local_val_' + str(f).replace('.', '_') + '.csv')
    df.loc[local_splits['test']].to_csv('splits/local_test_' + str(f).replace('.', '_') + '.csv')
